[#853] Bring in entries and modify places.
moveresults plugin renamed to modifydata with a more general usecase.
