+++ census/views/includes/dataviews/table_places.html
@@ -66,7 +66,7 @@
                   </ul>
                 {% endif %}
 
-                {% if is_index != "true" -%}
+                {% if not is_index -%}
                   {% if entry -%}
                     {% if entry.year == surveyYear -%}
                       {% set badgeClass = 'label-success' %}
