+++ index/src/about.md
@@ -1,4 +1,5 @@
 ---
 title: About
 layout: page.html
+breadcrumbTitle: About
 ---
